jwtauth/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
Django==6.0.1
django-allauth==65.14.0
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography-5==2.0.3
django-mailjet==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
mailjet-rest==1.5.1
MarkupSafe==3.0.3
mpmath==1.3.0
numpy==2.4.2
oauthlib==3.3.1
openai==2.16.0
packaging==26.0
proto-plus==1.27.1
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
setuptools==80.10.2
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
sslcommerz-client==0.6.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
psycopg2-binary
dj-database-url
dotenv
pgvector==0.4.2
google-genai
celery
redis
django-redis
django-celery-results
django-celery-beat
gunicorn==25.1.0